Refactor Desktop into a standalone Application
- Created `src/apps/desktop/DesktopApp.js` extending the base `Application` class.
- Moved taskbar initialization and wallpaper management into `DesktopApp`.
- Integrated `ZenDirectoryView` and `IconManager` for desktop icon rendering from `/C:/WINDOWS/Desktop`.
- Updated `ZenContextMenuBuilder` to launch `displayproperties` when right-clicking the desktop background.
- Simplified `src/main.js` boot sequence to launch the 'desktop' app.
- Removed obsolete `src/apps/zenexplorer/ZenDesktopApp.js`.
